Mark frameworks as safe for app extensions
Given that we're not using anything that can't be used in an app
extension, it really makes sense to restrict ourselves to APIs that can
be used in app extensions. This won't restrict usage for applications,
but will open up the ability to use Box (and anything that depends on
Box) inside extensions.